+++ app/components/ui/context-menu/index.js
@@ -1,4 +1,5 @@
 export { default as ContextMenu } from './ContextMenu.vue';
 export { default as ContextMenuContent } from './ContextMenuContent.vue';
 export { default as ContextMenuItem } from './ContextMenuItem.vue';
+export { default as ContextMenuSeparator } from './ContextMenuSeparator.vue';
 export { default as ContextMenuTrigger } from './ContextMenuTrigger.vue';